pyproject.toml
[project]
name = "bookleaf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "pandas>=2.3.3",
    "pymupdf>=1.27.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "streamlit>=1.54.0",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]
